refactor set min max only long
